Split up large files in plasmod.solver
